Rediseñando formularios.
- Se cambiaron los estilos de todos los formularios.
- Se creo un componente Login.
- Se creo el componente Register.
- Se optimizo la logica para el renderizado del componente Login y Register.
