+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import FinanceBoard from './components/pages/financeBoard'
-import Login from './components/pages/login'
+import Form from './components/pages/login'
 import { BrowserRouter as Router, Route } from 'react-router-dom'
 
 
@@ -10,7 +10,7 @@ const App = () => {
         <Router >
             <Route exact path="/" component={ FinanceBoard }/>
             <Route path="/board" component={ FinanceBoard } />
-            <Route path="/pruebas" component={ Login } />
+            <Route path="/login" component={ Form } />
         </Router>
     )
 }
