+++ src/components/pages/login/login.css
@@ -1,87 +1,7 @@
-.login {
-    background-color: rgb(34,34,34);
-    border: 2px solid rgb(41,41,41);
-    border-radius: 7px;
-    color: #fff;
+.container {
+    height: calc(100vh - 70px);
+    width: 100%;
     display: flex;
-    flex-direction: column;
-    line-height: 1.6;
-    margin: auto;
-    margin-top: 100px;
-    padding: 30px;
-    width: 400px;
-}
-
-.loginClosed {
-    display: none;
-}
-
-h2 {
-    font-weight: 400;
-    font-size: 30px;
-    margin: 20px 0px;
-}
-
-.loginInpt {
-    background-color: rgb(33,38,45);
-    border: 0.5px solid gray;
-    border-radius: 5px;
-    color: rgba(255, 255, 255, 0.425);
-    margin-bottom: 10px;
-    outline: none;
-    padding: 10px;
-    transition: 350ms;
-}
-
-.loginInpt:hover {
-    transform: scale(1.025);
-}
-
-.btnInit {
-    background-color: #348637;
-    border-radius: 10px;
-    border: none;
-    color: #fff;
-    font-size: 14px;
-    padding: 10px;
-    outline: none;
-    width: 200px;
-    margin: auto;
-}
-
-.btnInit:hover {
-    background-color: #2ea043;
-}
-
-.signup {
-    text-align: center;
-}
-
-.signup span {
-    color: #fff;
-    font-style: italic;
-    text-decoration: none;
-    margin-left: 10px;
-}
-
-.signup span:hover {
-    text-decoration: underline;
-}
-
-.signupForm {
-    display: none;
-}
-
-.signupFormOpen {
-    background-color: rgb(34,34,34);
-    border: 2px solid rgb(41,41,41);
-    border-radius: 7px;
-    color: #fff;
-    display: flex;
-    flex-direction: column;
-    line-height: 1.6;
-    margin: auto;
-    margin-top: 100px;
-    padding: 30px;
-    width: 400px;
+    justify-content: center;
+    align-items: center;
 }
